Individual user tied to the lab, they have their own progress. To do: show lab code on game pages, recolor and design profile and list pages, and already found reactant message working

+++ src/pages/profile/profile.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { NavController, NavParams, AlertController } from 'ionic-angular';
 import { AngularFireAuth } from 'angularfire2/auth';
 import { IndivLabPage } from '../indivlab/indivlab';
@@ -19,7 +19,9 @@ import { ActionSheetController } from 'ionic-angular';
 export class ProfilePage {
     email: string;
     newLab: Lab = {
+      purpose: "",
       entryCode: "",
+      sharedWith: [],
       name: "",
       originalCreator: "",
       combinationsIDs: [],
@@ -31,20 +33,22 @@ export class ProfilePage {
     helperText: string;
     users: any;
     myUser: User;
+    myUserIds: String[];
+    entryCodeAttempt: String = "";
     userFound: boolean = false;
+    currentLabs: Lab[] = [];
+
   constructor(private userService: UserService, public navParams: NavParams, private comboService: CombinationService, private materialService: MaterialService, private labService: LabService, public actionSheetCtrl: ActionSheetController, private fire: AngularFireAuth, public navCtrl: NavController, public alertCtrl: AlertController) {
+    this.myUser = this.navParams.data;
+    this.myUserIds = this.userService.getUserLabIds(this.myUser);
+    this.userFound = true;
+    console.log("current labs" + this.currentLabs);
     this.email = this.fire.auth.currentUser.email;
     this.userId = this.fire.auth.currentUser.uid;
-    this.myUser = this.navParams.data;
-    this.userFound = true;
-    console.log("MyUser after getting it from nav.Params.data (userName)" + this.myUser.userName);
     this.users = this.userService.getUsers(); // getting users in database
-    this.myUser.labIds = this.userService.getUserLabIds(this.myUser);
+    console.log("check show labs");
+    this.currentLabs = this.labService.getLabs(this.myUserIds);
   };
-  getMyLabs() {
-      console.log("this.myUser.labIds on page: " + this.myUser.labIds);
-      return this.labService.getLabs(this.myUser.labIds);
-  }
   // called to assign each lab its own entry code
   generateRandomLabCode(length: number) {
     let text: string = "";
@@ -54,6 +58,94 @@ export class ProfilePage {
       text = possibleValues.charAt(Math.floor(Math.random() * possibleValues.length)) + text;
     }
     return text;
+  }
+  joinLab() {
+    let labCodes = this.labService.getEntryCodes();
+    console.log("labCodes: " + labCodes);
+    if (labCodes.indexOf(this.entryCodeAttempt) !== -1) {
+      let labs = this.labService.getAllLabs();
+      let newLab = labs.find(lab => lab.entryCode == this.entryCodeAttempt);
+      if (this.myUserIds.indexOf(newLab.id) ==  -1) {
+        this.myUser.myLabs.push({labId: newLab.id, isFoundIds:[]});
+        this.userService.updateUser(this.myUser);
+      }
+    }
+  }
+  addPurposeConfirm(lab: Lab) {
+    let alert = this.alertCtrl.create({
+      title: 'Edit Purpose',
+      message: 'What is the purpose of this lab?',
+      inputs: [
+        {
+          name: 'Purpose',
+          placeholder: 'The purpose of this lab is to ...'
+        },
+      ],
+      buttons: [
+        {
+          text: 'Cancel',
+          role: 'cancel',
+          handler: () => {
+            console.log('Cancel clicked');
+          }
+        },
+        {
+          text: 'Save',
+          handler: (purpose: string) => {
+            lab.purpose = purpose;
+            this.labService.updateLab(lab);
+          }
+        }
+      ]
+    });
+    alert.present();
+  }
+  addInvitationConfirm(lab: Lab) {
+    let alert = this.alertCtrl.create({
+      title: 'Invite Others',
+      message: 'Who would you like to invite?',
+      inputs: [
+        {
+          name: 'Email',
+          placeholder: 'labplus@labplus.com'
+        },
+      ],
+      buttons: [
+        {
+          text: 'Cancel',
+          role: 'cancel',
+          handler: () => {
+            console.log('Cancel clicked');
+          }
+        },
+        {
+          text: 'Invite',
+          handler: otherUserName => { //just collecting/comparing username for now
+            let userNames = this.userService.getUserNames();
+            if (userNames.indexOf(otherUserName.Email) !== -1) { // make sure other user exists
+              let otherUser = this.users.find(user => user.userName == otherUserName.Email);
+              // make sure not already in lab.sharedWith
+              if (lab.sharedWith.indexOf(otherUser.userId) == -1) { // make sure it is not already in list
+                lab.sharedWith.push(otherUser.userId);
+                this.labService.updateLab(lab);
+              } this.alert("You have already joined this lab.");
+              // you already have this lab
+            } this.alert("This user is not registered with LabPlus");// else the user is not registered ( but alert is not working right now )
+              // add message to other user or send to their notifications later
+          }
+        }
+      ]
+    });
+    alert.present();
+  }
+  alert(message: String) {
+    alert(message);
+  }
+  addInvitation(lab: Lab) {
+    this.addInvitationConfirm(lab);
+  }
+  addPurpose(lab: Lab) {
+    this.addPurposeConfirm(lab);
   }
   deleteLab(lab: Lab) {
     let currentCombos = this.comboService.getCombos(lab.combinationsIDs);
@@ -72,35 +164,43 @@ export class ProfilePage {
       this.materialService.deleteMaterial(material);
     })
     
-    // delete individual combination
-    let labIdIndex = this.myUser.labIds.indexOf(lab.id);
-    this.myUser.labIds.splice(labIdIndex, labIdIndex+1);
-    this.labService.deleteLab(lab);
+    // delete individual lab
+    let i = 0;
+    this.myUser.myLabs.forEach((myLab) => {
+      if (myLab.labId == lab.id) {
+        this.myUser.myLabs.splice(i,i+1);
+        this.labService.deleteLab(lab);
+      }
+      i++;
+    })
   }
-  goToIndivLabPage(lab: Lab) {
-    this.navCtrl.push(IndivLabPage, lab);
+  goToIndivLabPage(lab: Lab) { // lab from lab db 
+    this.navCtrl.push(IndivLabPage,{data1: lab, data2: this.myUser}); 
   }
   // method to be used after lab created (in onSubmit() function)
   onSubmit() {
     this.newLab.entryCode = this.generateRandomLabCode(8); // generating specific entry code to lab --> display later in html 
     this.newLab.originalCreator = this.userId; // assigning the id to original Creator
     this.newLab.combinationsIDs = [];
+    this.newLab.sharedWith = [];
     this.newLab.materialsIDs = [];
     this.newLab.isFinished = false;
-    this.newLab.isFoundIDs = [];
+    this.newLab.purpose = "";
     if((this.newLab.name != "") && (this.newLab.originalCreator != "") && (this.newLab.entryCode != "")) {
 
       let ref = this.labService.addLab(this.newLab);
-      ref.then(c => { this.myUser.labIds.push(c.id);
+      ref.then(c => {
+        this.myUser.myLabs.push({labId: c.id.toString(), isFoundIds: []});        
         this.userService.updateUser(this.myUser);
       });
       this.newLab.name = "";
       this.newLab.entryCode = "";
+      this.newLab.sharedWith = [];
+      this.newLab.purpose = "";
       this.newLab.combinationsIDs = [];
       this.newLab.materialsIDs = [];
       this.newLab.originalCreator = "";
       this.newLab.isFinished = false;
-      this.newLab.isFoundIDs = [];
     }
     else {
       // ~~ERROR~~ need to change if statement, doesn't work if the firnla one is put in before the first one. As they type into fields, the fields should be updating. (two way data binding)
